Finish working prototype
It's a mess. But the main logic where chunks of one waveform are correlated with chunks of a bunch of different other waveforms, and then the chunks of the second set with the best correlation to the first wav are concatenated into an output wav file, is complete.

There are a lot of holes with the code. It's not samplerate dependent, but it kinda assumes that sample rates match, furthermore, I've only tested with stereo data, I'm sure the script completely breaks when 1-channel and 2-channel wavs are both fed to it.