Test vector normalization by maximum value
Test that arrays are normalized as expected and that data types are correct.

Arrays, representing wavs, are processed on a by-channel basis. That is, the left channel of wav 2 is normalized to have the same maximum as the left channel of wav 1 regardless of whatever is going on with the right channel of wav 1.

Also ensure that the datatype of the resulting array is equivalent to the input array. The division required by normalization will initiallly cast the result to a float type that needs to be downcast to an integer type to be compatible with wavio.