Improve normalization method, add get_valid_wavs()
Ensures that the datatype of the resulting array within the normalization method is equivalent to the input array. The division required by normalization will initiallly cast the result to a float type that needs to be downcast to an integer type to be compatible with wavio.

Also encapsulates code that looks through a directory, checks all files to see if they are valid wavs, and then stored full filepaths to valid wavs in a list, into a reusable function.